+++ CHANGELOG.md
@@ -15,6 +15,9 @@ OR
 
 Changelog
 =========
+## 1.1.1 - 19 September 2019
+
+- Fixed [Issue #5](https://github.com/timothycrosley/hypothesis-auto/issues/5): typo in Scenario class name
 
 ## 1.1.0 - 18 September 2019
 
@@ -22,7 +25,7 @@ Changelog
     This release contains breaking changes. In particular to how parameters are passed in to testing functions.
 
 - `_prefixing` for parameters was replaced with `suffixing_` to avoid private implication while still avoiding parameter name conflicts.
-- Custom `auto_verify_` function now must take a single [`Scenerio`](https://timothycrosley.github.io/hypothesis-auto/reference/hypothesis_auto/tester/#scenerio) object.
+- Custom `auto_verify_` function now must take a single [`Scenario`](https://timothycrosley.github.io/hypothesis-auto/reference/hypothesis_auto/tester/#scenario) object.
 
 ## 1.0.0 - 17 September 2019
 - Initial Release.